# PERF: eliminate unnecessary ravel for median with order="F", speedup of up to 2x